Discard commented out code
We can trust that the code we might want to comments out will be available to us.

If it was once working code then it lives on in our source control history.
If it is not ready for primetime we can leave it in a branch that hasn't merged to master.